+++ tests/testthat/test-tidyeval-across.R
@@ -365,6 +365,29 @@ test_that("across() searches for list in environment", {
   )
 })
 
+test_that("across() handles cur_column()", {
+  lf <- lazy_frame(a = 1, b = 2)
+
+  expect_equal(
+    capture_across(lf, across(a:b, ~ paste0(.x, cur_column()))),
+    exprs(a = paste0(a, "a"), b = paste0(b, "b"))
+  )
+
+  expect_equal(
+    capture_across(lf, across(a:b, function(x) paste0(x, cur_column()))),
+    exprs(a = paste0(a, "a"), b = paste0(b, "b"))
+  )
+})
+
+test_that("across() errors if named", {
+  lf <- lazy_frame(a = 1, b = 2)
+
+  expect_snapshot({
+    (expect_error(mutate(lf, x = across())))
+    (expect_error(group_by(lf, x = across())))
+  })
+})
+
 
 # if_all ------------------------------------------------------------------