+++ tests/testthat/test-verb-compute.R
@@ -96,6 +96,14 @@ test_that("compute can handle schema", {
   )
 })
 
+test_that("collect() handles DBI error", {
+  mf <- memdb_frame(x = 1)
+  expect_snapshot(
+    (expect_error(mf %>% mutate(a = sql("invalid sql")) %>% collect())),
+    transform = snap_transform_dbi
+  )
+})
+
 # ops ---------------------------------------------------------------------
 
 test_that("sorting preserved across compute and collapse", {